+++ assets/css/tailwind.css
@@ -282,3 +282,19 @@ body {
 .debug {
   @apply !border-2 !border-red-500;
 }
+
+.grid-container {
+  display: grid;
+  grid-template-columns: repeat(7, 1fr);
+  grid-template-rows: 0.5fr 0.3fr 2.4fr 1fr 2.3fr 0.4fr 0.5fr;
+  gap: 0px 0px;
+}
+.header-container {
+  grid-area: 1 / 1 / 2 / 8;
+}
+.main-content-container {
+  grid-area: 3 / 1 / 6 / 8;
+}
+.footer-container {
+  grid-area: 7 / 1 / 8 / 8;
+}
