+++ components/ui/UiFooter.vue
@@ -1,6 +1,6 @@
 <template>
   <footer
-    class="relative z-10 flex w-full items-center justify-between px-8 pb-4"
+    class="footer-container relative z-10 flex w-full items-center justify-between px-8 pb-4"
   >
     <nuxt-link
       to="/settings"
